minor changes
changed iterators to use pointers instead of references so that we are now able to assign iterators to one another after their initialization. Also added a check in operator == for both DllIterator and ConstDllIterator to verify whether list == rhs.list. This fixes the situation where, if we have two empty lists and we initialize two iterators one pointing to the beginning of the first list and the other to the beginning of the second list, the operator == would incorrectly return true